Rename and move transport participant factory
