+++ dds/src/transport/mod.rs
@@ -1,3 +1,4 @@
+pub mod factory;
 pub mod history_cache;
 pub mod participant;
 pub mod reader;
